refactor(managed-agent-runtime): extract SudoCodeSpawnAdapter to thin crate (PR-C v1 follow-up)
Plan PR-C v1 envisioned the cross-repo runtime-body adapter as a
shared rlib both binary edges (Python wheel cdylib + cluster binary)
depend on. The earlier commit `7db5517be` parked the adapter
directly in nexus-cdylib — that left the cluster binary without
runtime-body parity (cluster deployments would have shipped
`install_managed_agent` with no spawn provider, procfs-only).

This commit lifts the adapter into a dedicated `rust/managed-agent-runtime/`
workspace member so:

  - Both `nexus-cdylib` (Python wheel boot) and `profiles/cluster`
    (Rust binary boot) call the same canonical entry point —
    `managed_agent_runtime::install_managed_agent_with_sudocode_spawn(kernel)`.
  - The cross-repo `sudocode-runtime` git-dep lives once in
    `managed-agent-runtime/Cargo.toml`, not duplicated at each
    binary edge.
  - Future runtime adapters (a different LLM-driven loop, a
    headless runtime, …) join this rlib as additional
    `SpawnTask<Kernel>` impls and are immediately available to
    every binary edge.
  - Slim deployments that don't ship a runtime body (future
    raft-witness, WASM) skip the dep on this rlib and call
    `services::managed_agent::install_managed_agent` directly.

## Cluster binary parity

`profiles/cluster::run_daemon` now calls
`install_managed_agent_with_sudocode_spawn(&kernel)` after the
root-fs mount succeeds, before the topology apply loop spins up.
Cluster-binary deployments ship the same managed-agent runtime
body Python-wheel deployments do; no asymmetric "install path
depends on whether you booted via Python or pure Rust".

## Trait surface unchanged

`services::managed_agent::SpawnTask<K>` + `SpawnHandle` traits stay
in services (the framework boundary). Only the **concrete impl**
that wraps `sudocode_runtime::spawn_task::spawn_task::<Kernel>(...)`
moves out, into the new rlib that's allowed to carry the
cross-repo coupling.

Test gates: `cargo check --workspace` clean (cdylib + cluster
binary + managed-agent-runtime + sudocode-runtime via cross-repo
git-dep + every existing rlib).
