+++ rust/managed-agent-runtime/Cargo.toml
@@ -1,0 +1,28 @@
+[package]
+name = "managed-agent-runtime"
+version = "0.10.0"
+edition = "2021"
+description = "Cross-repo runtime-body adapters for `services::managed_agent`. Wraps third-party runtime crates (today: sudocode-runtime `spawn_task`) as concrete `services::managed_agent::SpawnTask<Kernel>` impls. Lives in its own rlib so both the Python wheel cdylib and the cluster binary can wire the same managed-agent runtime body without duplicating adapter code."
+
+[lib]
+name = "managed_agent_runtime"
+crate-type = ["rlib"]
+
+[dependencies]
+# Reach `services::managed_agent::{SpawnTask, SpawnHandle,
+# install_managed_agent_with_spawn}` to register the adapter at boot.
+services = { workspace = true, features = ["service-managed-agent"] }
+# Reach `KernelAbi` (the trait the SpawnTask impl is generic over) +
+# `AgentDescriptor` types. `default-features = false` keeps the
+# kernel runtime defaults off in this rlib — same convention every
+# peer rlib that depends on kernel uses.
+kernel = { workspace = true }
+# Sudo-code runtime crate — provides `runtime::spawn_task<K: KernelAbi>`
+# that the adapter wraps. Cross-repo coupling lives **here** in this
+# rlib (not in services or in a binary edge), so binary edges
+# (nexus-cdylib, profiles/cluster) get one canonical wiring entry.
+#
+# Lockstep with sudocode PR sudoprivacy/sudocode#99
+# (`feat/spawn-task-kernel-abi-v1`, rev `2b67d5b`). Bump rev pin in
+# this file when sudocode merges to main.
+sudocode-runtime = { package = "runtime", git = "https://github.com/sudoprivacy/sudocode", rev = "2b67d5b452ea69b8ea4aba1fca3fa73588df60da" }
